extect NewCidForTestGetter to testing utils
